backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Async and WebSockets
channels==4.0.0
channels-redis==4.1.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Authentication and Security
cryptography==41.0.7
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# File Processing
Pillow==10.1.0
python-multipart==0.0.6

# NLP and ML
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2

# FHIR and Healthcare
fhir.resources==7.0.2
hl7apy==1.3.4

# WebRTC and Video
aiortc==1.6.0

# Notifications
firebase-admin==6.2.0
twilio==8.10.3

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Utilities
pytz==2023.3
phonenumbers==8.13.26
qrcode==7.4.2

# API Documentation
drf-spectacular==0.26.5

# Production WSGI Server
gunicorn==21.2.0